SEARCH CLASS: Added a getCurrentFilters() Added a searchForRefNum(int refNum) method that will find a course with the same reference number Made it so that the search class requires a Semester in its constructor. it will only allow you to search for classes that match the semester entered.
SESSION CLASS:
Moved where we make the searchBox
Changed filter() to print all filters applied.
I added a stub method called removeSpecificFilter().
Made an option to add a course in scheduleMenu(). Also made it so that it throws Excetion in method signature.
Made the addByReference() method that allows a user to add a course to their schedule directly by inputting the reference number.
Added a addCourseMenu() method that the user will interact with.